Moved io package to util
