+++ nifi-nar-bundles/nifi-hadoop-bundle/nifi-hdfs-processors/src/main/java/org/apache/nifi/processors/hadoop/util/ByteFilteringOutputStream.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.nifi.processors.hadoop.util.io;
+package org.apache.nifi.processors.hadoop.util;
 
 import java.io.FilterOutputStream;
 import java.io.IOException;
